+++ core/src/main/java/react4j/arez/ReactArezComponent.java
@@ -1,6 +1,5 @@
 package react4j.arez;
 
-import arez.Arez;
 import arez.ArezContext;
 import arez.Disposable;
 import arez.Observer;
@@ -10,15 +9,13 @@ import arez.annotations.ComponentNameRef;
 import arez.annotations.ContextRef;
 import arez.annotations.ObserverRef;
 import arez.annotations.PreDispose;
-import arez.spy.ObservableValueInfo;
 import arez.spy.ObserverInfo;
 import elemental2.promise.Promise;
-import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
 import react4j.Component;
+import react4j.internal.arez.IntrospectUtil;
 
 /**
  * A base class for all Arez enabled components. This class makes the component
@@ -176,80 +173,6 @@ public abstract class ReactArezComponent
 
     // Collect existing dependencies as state
     final ObserverInfo observerInfo = getContext().getSpy().asObserverInfo( getRenderObserver() );
-    observerInfo.getDependencies().forEach( d -> newState.set( d.getName(), getValue( d ) ) );
-  }
-
-  /**
-   * Return the value for specified observable.
-   * Exceptions are caught and types are converted to strings using {@link Object#toString()}
-   *
-   * @param observableInfo the observable.
-   * @return the value as a string.
-   */
-  @SuppressWarnings( "UnnecessaryUnboxing" )
-  @Nullable
-  private Object getValue( @Nonnull final ObservableValueInfo observableInfo )
-  {
-    try
-    {
-      if ( Arez.arePropertyIntrospectorsEnabled() && observableInfo.hasAccessor() )
-      {
-        // Consider unwrapping collections and potentially serializing Arez entities so they are presented correctly in DevTools
-        final Object value = observableInfo.getValue();
-        if ( null == value )
-        {
-          return null;
-        }
-        else if ( value instanceof Enum )
-        {
-          return ( (Enum) value ).name();
-        }
-        else if ( value instanceof Integer )
-        {
-          return Js.asAny( ( (Integer) value ).intValue() );
-        }
-        else if ( value instanceof Boolean )
-        {
-          return Js.asAny( ( (Boolean) value ).booleanValue() );
-        }
-        else if ( value instanceof Long )
-        {
-          return Js.asAny( ( (Long) value ).doubleValue() );
-        }
-        else if ( value instanceof Float )
-        {
-          return Js.asAny( ( (Float) value ).doubleValue() );
-        }
-        else if ( value instanceof Short )
-        {
-          return Js.asAny( ( (Short) value ).intValue() );
-        }
-        else if ( value instanceof Byte )
-        {
-          return Js.asAny( ( (Byte) value ).intValue() );
-        }
-        else if ( value instanceof Character )
-        {
-          return value.toString();
-        }
-        else if ( value instanceof Stream )
-        {
-          // Streams are new instances every time so render them as strings to avoid infinite loops.
-          return "<Stream>";
-        }
-        else
-        {
-          return value;
-        }
-      }
-      else
-      {
-        return "?";
-      }
-    }
-    catch ( final Throwable throwable )
-    {
-      return throwable;
-    }
+    observerInfo.getDependencies().forEach( d -> newState.set( d.getName(), IntrospectUtil.getValue( d ) ) );
   }
 }