+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * **\[core\]** Merge the `react4j-arez` artifact into `react4j-core` as the `react4j-core` artifact has never
   been used without also using `react4j-arez` in any production application.
+* **\[core\]** Move the `ObservableValue` introspection utilities from `ReactArezComponent` into a new utility
+  class `react4j.internal.arez.IntrospectUtil` as the code is not expected to be used by end-user code.
 
 ### [v0.113](https://github.com/react4j/react4j/tree/v0.113) (2019-01-21)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.112...v0.113)